requirements.txt
bottle==0.13.2
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pillow==11.2.1
qrcode==8.1
requests==2.32.3
setuptools==78.1.0
urllib3==2.4.0
zbarlight==3.0
